working julia <-> python test
